requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.2
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-multipart>=0.0.20
python-dotenv>=1.1.0
langchain>=0.2.10
langchain-community>=0.2.10
langchain-core>=0.1.0
langchain-google-genai>=0.1.15
google-generativeai>=0.8.5
pypdf>=5.5.0 
PyMuPDF>=1.23.0
pydantic[email]>=2.7.1
